Teacher Function: - Creata la funzione che restituisce gli eventi dell'università attualmente solo nei servizi del docente (eventualmente è possibile visaualizzarli anche nei servizi studente). - Bug fix frontend nel ruolo studente.

+++ appuniparthenope/lib/model/studentService/calendar_data.dart
@@ -66,9 +66,9 @@ final descriptionValues = EnumValues(
 
 class Room {
   final String name;
-  final int capacity;
+  final double capacity;
   final String description;
-  final int availability;
+  final double availability;
 
   Room({
     required this.name,
